Generate runtime STUB traps and drop /FORCE from the runnable links
Wiz8Runtime.exe and Wiz8RuntimeTest.exe linked /FORCE:UNRESOLVED, so every
unrecovered call jumped into the PE header and destroyed the caller's stack
before faulting. The recovered-object unresolved frontier now feeds a
build-time generator: wiz8 generate runtime-stubs resolves each exact
decorated symbol against address-qualified names, declaration address
comments, and one read-only Ghidra session, then writes runtime_stubs.cpp,
runtime_stubs.json and a generated COFF thunk object. VC6's LINK has no
/alternatename, so the object defines each symbol as a jmp into a unique
non-returning trap that prints WIZ8_RUNTIME_STUB and breaks, leaving the
caller's stack intact. The comparison image keeps /FORCE and never links the
object; the runnable links succeed without it, with the CRT's absolute
__except_list supplied alongside.

The generator refuses to stub data, RTTI or vtable storage, and fails with
the exact symbol when an unresolved spelling maps to an already recovered
retail address. That exposed and fixed one stale caller identity
(Function500460 -> AbsorbMissileDamage00500460), one misnamed SGP call
(RefreshSlfArchives -> ReopenCDLibraries), five missing data globals, and
three SurRender declaration/import mismatches (srFog destructor visibility,
getPolyShader layer type, missing srModelInstance/srMaterialIFace imports).

The semantic scenario now waits for the menu playlist before observing it,
which removes a startup race.

+++ include/surrender/srMeshModel.h
@@ -119,7 +119,7 @@ public:
     srTextureIFace* getTexture(long polygon, long layer) const;
     void setMaterial(srMaterialIFace* material, long polygon, e_side side);
     void setTexture(srTextureIFace* texture, long polygon, long layer);
-    srShader* getPolyShader(long polygon, long layer);
+    srShader* getPolyShader(long polygon, int layer);
     srShader getShader(long polygon) const;
     void setShader(srShader shader, long pass);
     void setUVCount(long count);
